requirements.txt
# Snooker Detection Pro - Complete Requirements
# Core Dependencies

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Computer Vision & ML
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
scipy==1.11.3

# Visualization & Analysis
matplotlib==3.7.2
scikit-learn==1.3.0

# File Handling
Pillow==10.0.1
python-multipart==0.0.6

# Utilities
PyYAML==6.0.1
tqdm==4.66.1
filterpy==1.4.5

# Testing (Optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Production Deployment (Optional)
gunicorn==21.2.0